+++ app/src/main/java/app/aryan447/mpvium/preferences/AppearancePreferences.kt
@@ -59,7 +59,7 @@ class AppearancePreferences(
   val bottomRightControls =
     preferenceStore.getString(
       "bottom_right_controls",
-      "PICTURE_IN_PICTURE,ASPECT_RATIO",
+      "PICTURE_IN_PICTURE,ASPECT_RATIO,HOLD_GESTURE_TOGGLE",
     )
 
   val bottomLeftControls =